total size of tensors is size_t
total size of tensors is size_t
total number of parameters is int64_t

+++ include/llama.h
@@ -469,10 +469,10 @@ extern "C" {
     LLAMA_API int32_t llama_model_desc(const struct llama_model * model, char * buf, size_t buf_size);
 
     // Returns the total size of all the tensors in the model in bytes
-    LLAMA_API uint64_t llama_model_size(const struct llama_model * model);
+    LLAMA_API size_t llama_model_size(const struct llama_model * model);
 
     // Returns the total number of parameters in the model
-    LLAMA_API uint64_t llama_model_n_params(const struct llama_model * model);
+    LLAMA_API int64_t llama_model_n_params(const struct llama_model * model);
 
     // Get a llama model tensor
     LLAMA_API struct ggml_tensor * llama_get_model_tensor(struct llama_model * model, const char * name);
